requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
Command==0.1.0
cryptography==40.0.2
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.1
django-allauth==0.54.0
django-multiselectfield==0.1.12
django-phone-field==1.8.1
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-user-language-middleware==0.0.3
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
filelock==3.9.0
idna==3.4
oauthlib==3.2.2
pbr==5.11.1
phonenumbers==8.13.11
Pillow==9.4.0
platformdirs==3.1.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
stevedore==5.0.0
tomli==2.0.1
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.20.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.7
